infra/local/requirements.txt
azure-storage-blob==12.20.0
azure-storage-queue==12.9.0